requirements.txt
pytorch-lightning
l5kit
numpy
matplotlib
seaborn
pandas
svgwrite
cairosvg
IPython
six>=1.12.0
tensorflow
tensorboardX
pillow
svglib
tqdm
jupyter
scikit-image
pandas
networkx
moviepy
numba
sklearn
umap-learn
umap-learn
shapely
kivy